Add CRUD

+++ src/main/java/ru/evgkit/giflib/service/CategoryServiceImpl.java
@@ -29,6 +29,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Category category) {
+        if (!category.getGifs().isEmpty()) {
+            throw new CategoryNotEmptyException();
+        }
+
         categoryDao.delete(category);
     }
 }
